Refactoring and adding new files
