--- src/main/java/com/foodiesfinds/recipe_service/dto/TagDTO.java
@@ -1,12 +1,0 @@
-package com.foodiesfinds.recipe_service.dto;
-
-import com.foodiesfinds.recipe_service.entity.RecipeTagId;
-import lombok.Data;
-
-@Data
-public class TagDTO {
-
-  Long tagId;
-  String tagName;
-
-}
