--- src/main/java/com/foodiesfinds/recipe_service/dto/UnitDTO.java
@@ -1,16 +1,0 @@
-package com.foodiesfinds.recipe_service.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class UnitDTO {
-
-  private Long id;
-  private String name;
-  private String abbreviation;
-
-}
