+++ src/main/java/com/foodiesfinds/recipe_service/dto/UnitResponseDTO.java
@@ -1,20 +1,18 @@
 package com.foodiesfinds.recipe_service.dto;
 
-import java.math.BigDecimal;
+import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-
 @Data
-@NoArgsConstructor
 @AllArgsConstructor
-public class RecipeIngredientSaveDTO {
+@NoArgsConstructor
+public class UnitDTO {
+
+  @NotNull(message = "Unit ID is required")
   private Long id;
   private String name;
-  private Long unitId;
-  private String unitName;
   private String abbreviation;
-  private BigDecimal quantity;
-  private String notes;
+
 }
