+++ src/main/java/com/foodiesfinds/recipe_service/dto/base/BaseInstructionStepDTO.java
@@ -1,0 +1,4 @@
+package com.foodiesfinds.recipe_service.dto.base;
+
+public class BaseInstructionStepDTO {
+}
