+++ src/main/java/com/foodiesfinds/recipe_service/dto/ingredient/RecipeIngredientResponseDTO.java
@@ -1,0 +1,14 @@
+package com.foodiesfinds.recipe_service.dto.ingredient;
+
+import com.foodiesfinds.recipe_service.dto.base.BaseRecipeIngredientDTO;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class RecipeIngredientDTO extends BaseRecipeIngredientDTO {
+  private Long recipeIngredientId;
+}
