+++ src/main/java/com/foodiesfinds/recipe_service/dto/instruction/InstructionStepResponseDTO.java
@@ -1,0 +1,13 @@
+package com.foodiesfinds.recipe_service.dto.instruction;
+
+import com.foodiesfinds.recipe_service.dto.base.BaseInstructionStepDTO;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class InstructionStepDTO extends BaseInstructionStepDTO {
+  private Long stepId;
+}
