+++ src/main/java/com/foodiesfinds/recipe_service/dto/instruction/InstructionStepSaveDTO.java
@@ -1,5 +1,6 @@
 package com.foodiesfinds.recipe_service.dto;
 
+import com.foodiesfinds.recipe_service.dto.base.BaseInstructionStepDTO;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -7,7 +8,6 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class RecipeTagSaveDTO {
-  private Long id;
-  String name;
+public class InstructionStepSaveDTO extends BaseInstructionStepDTO {
+
 }
