+++ src/main/java/com/foodiesfinds/recipe_service/dto/recipe/RecipeResponseDTO.java
@@ -1,5 +1,9 @@
-package com.foodiesfinds.recipe_service.dto;
+package com.foodiesfinds.recipe_service.dto.recipe;
 
+import com.foodiesfinds.recipe_service.dto.instruction.InstructionStepDTO;
+import com.foodiesfinds.recipe_service.dto.instruction.RecipeIngredientDTO;
+import com.foodiesfinds.recipe_service.dto.tag.RecipeTagDTO;
+import com.foodiesfinds.recipe_service.dto.base.BaseRecipeDTO;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
@@ -10,7 +14,7 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class RecipeDTO extends BaseRecipeDTO<
+public class RecipeResponseDTO extends BaseRecipeDTO<
     RecipeIngredientDTO,
     RecipeTagDTO,
     InstructionStepDTO
