+++ src/main/java/com/foodiesfinds/recipe_service/dto/recipe/RecipeSaveDTO.java
@@ -1,5 +1,7 @@
 package com.foodiesfinds.recipe_service.dto;
 
+import com.foodiesfinds.recipe_service.dto.base.BaseRecipeDTO;
+
 public class RecipeSaveDTO extends BaseRecipeDTO<
     RecipeIngredientSaveDTO,
     RecipeTagSaveDTO,
