+++ src/main/java/com/foodiesfinds/recipe_service/dto/tag/RecipeTagResponseDTO.java
@@ -1,0 +1,13 @@
+package com.foodiesfinds.recipe_service.dto.tag;
+
+import com.foodiesfinds.recipe_service.dto.base.BaseRecipeTagDTO;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class RecipeTagDTO extends BaseRecipeTagDTO {
+  private Long recipeTagId;
+}
