+++ src/main/java/com/foodiesfinds/recipe_service/dto/tag/TagResponseDTO.java
@@ -1,0 +1,9 @@
+package com.foodiesfinds.recipe_service.dto.tag;
+
+import lombok.Data;
+
+@Data
+public class TagDTO {
+  Long tagId;
+  String tagName;
+}
