+++ src/main/java/com/foodiesfinds/recipe_service/repository/UnitRepository.java
@@ -1,0 +1,4 @@
+package com.foodiesfinds.recipe_service.repository;
+
+public interface UnitRepository {
+}
