+++ src/main/java/com/foodiesfinds/recipe_service/service/UnitService.java
@@ -1,0 +1,4 @@
+package com.foodiesfinds.recipe_service.service.implementation;
+
+public interface UnitService {
+}
