+++ src/main/java/com/foodiesfinds/recipe_service/service/implementation/UnitServiceImpl.java
@@ -1,0 +1,4 @@
+package com.foodiesfinds.recipe_service.service;
+
+public class UnitServiceImpl {
+}
